requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.4
django-allauth==0.56.1
django-crispy-forms==2.1
django-registration-redux==2.13
django-shopping-cart==0.1
dnspython==2.4.2
email-validator==2.0.0.post2
filelock==3.12.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
peppercorn==0.6
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sampleproject==3.0.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
stripe==7.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.4
Werkzeug==2.3.7
WTForms==3.0.1
